+++ packages/config-controller/package.json
@@ -16,7 +16,7 @@
     "build": "tsc --build ./tsconfig.json",
     "build:watch": "nodemon --watch \"./src/**/*.*\" --exec npm run build",
     "test": "jest --testTimeout=25000",
-    "test:watch": "jest --watchAll --testTimeout=25000"
+    "test:watch": "jest --watch --testTimeout=25000"
   },
   "repository": {
     "directory": "packages/config-controller",
