+++ packages/config-controller/src/index.ts
@@ -1,10 +1,10 @@
 /*
- *
- * @author Sawyer Cutler
  * @license MIT
- * @copyright 2022 Dirt Road Dev
- * @package @skaleproject/etherbase
+ * @copyright (c) 2022 Dirt Road Dev
+ * @package @skaleproject/config-controller
  *
-**/
+ * @file index.ts
+ * @author Sawyer Cutler
+*/
 
-export * from "./config_controller";
+export * from "./contract";
